+++ packages/grid/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @illa-design/list
 
+## 0.0.8
+
+### Patch Changes
+
+- fix bugs
+
+- Updated dependencies []:
+  - @illa-design/system@0.0.8
+  - @illa-design/theme@0.0.8
+
+## 0.0.7
+
+### Patch Changes
+
+- bugfix
+
+- Updated dependencies []:
+  - @illa-design/system@0.0.7
+  - @illa-design/theme@0.0.7
+
 ## 0.0.6
 
 ### Patch Changes