+++ packages/popover/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @illa-design/popover
 
+## 0.0.8
+
+### Patch Changes
+
+- fix bugs
+
+- Updated dependencies []:
+  - @illa-design/link@0.0.8
+  - @illa-design/system@0.0.8
+  - @illa-design/theme@0.0.8
+  - @illa-design/trigger@0.0.8
+  - @illa-design/typography@0.0.8
+
+## 0.0.7
+
+### Patch Changes
+
+- bugfix
+
+- Updated dependencies []:
+  - @illa-design/link@0.0.7
+  - @illa-design/system@0.0.7
+  - @illa-design/theme@0.0.7
+  - @illa-design/trigger@0.0.7
+  - @illa-design/typography@0.0.7
+
 ## 0.0.6
 
 ### Patch Changes