+++ packages/system/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @illa-design/system
 
+## 0.0.8
+
+### Patch Changes
+
+- fix bugs
+
+## 0.0.7
+
+### Patch Changes
+
+- bugfix
+
 ## 0.0.6
 
 ### Patch Changes